Keep Apple target detection in planning
Remove the preparation detection mode and move Xcode/runtime/device selection into the matrix planner, leaving preparation responsible only for restoring tools and resetting output.


Copilot-Session: c25c2c8a-2652-40ae-8fcc-c232414e9eb7
