+++ .agents/skills/release-testing/scripts/tests/test_prepare_test_run.py
@@ -17,21 +17,6 @@ sys.modules[SPEC.name] = preparer
 SPEC.loader.exec_module(preparer)
 
 
-def simulator(version: str, device: str) -> dict:
-    return {
-        "isAvailable": True,
-        "deviceType": {
-            "name": device,
-            "productFamily": "iPhone",
-        },
-        "runtime": {
-            "name": f"iOS {version}",
-            "version": version,
-            "isAvailable": True,
-        },
-    }
-
-
 class PrepareTestRunTests(unittest.TestCase):
     def test_reset_output_removes_only_integration_artifacts(self):
         with tempfile.TemporaryDirectory() as temporary:
@@ -61,42 +46,6 @@ class PrepareTestRunTests(unittest.TestCase):
             '["dotnet", "tool", "restore"]',
             source,
         )
-        self.assertIn('"xcode",', source)
-
-    def test_xcode_26_selects_ios_15_and_26(self):
-        targets = preparer.select_apple_targets(
-            "26.6",
-            "/Applications/Xcode-26.6.0.app",
-            [
-                simulator("15.8", "iPhone 13 Pro"),
-                simulator("15.0", "iPhone 13"),
-                simulator("26.0", "iPhone 16"),
-                simulator("26.5", "iPhone 17"),
-            ],
-        )
-
-        self.assertEqual(targets["minimum"]["version"], "15.0")
-        self.assertEqual(targets["minimum"]["device"], "iPhone 13")
-        self.assertEqual(targets["maximum"]["version"], "26.5")
-        self.assertEqual(targets["maximum"]["device"], "iPhone 17")
-
-    def test_xcode_27_selects_ios_18_and_26(self):
-        targets = preparer.select_apple_targets(
-            "27.0",
-            "/Applications/Xcode.app",
-            [
-                simulator("16.2", "iPhone 14"),
-                simulator("18.2", "iPhone 16"),
-                simulator("18.5", "iPhone 16 Pro"),
-                simulator("26.5", "iPhone 17"),
-                simulator("27.0", "iPhone 17 Pro"),
-            ],
-        )
-
-        self.assertEqual(targets["minimum"]["version"], "18.2")
-        self.assertEqual(targets["minimum"]["device"], "iPhone 16")
-        self.assertEqual(targets["maximum"]["version"], "26.5")
-        self.assertEqual(targets["maximum"]["device"], "iPhone 17")
 
     def test_help_is_inert(self):
         result = subprocess.run(
